requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
apturl==0.5.2
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.4
Brlapi==0.8.3
certifi==2025.4.26
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
liewa==3.2
louis==3.20.0
macaroonbakery==1.3.1
magic-filter==1.0.12
matplotlib==3.10.3
more-itertools==8.10.0
multidict==6.4.3
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
opencv-python==4.11.0.86
packaging==25.0
paho-mqtt==2.1.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.7
propcache==0.3.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.7
ssh-import-id==5.11
systemd-python==234
typing-inspection==0.4.0
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.30.0
wadllib==1.3.6
xdg==5
xkit==0.0.0
yarl==1.20.0
zipp==1.0.0
